// i-guest-house-ui/src/app/services/models/booking-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface BookingRequest {
  bookedDate?: string;
  bookedToId: string;
  bookingId?: number;
  checkInDate: string;
  checkOutDate: string;
  guestName: string;
  roomNumber?: number;
  roomType: string;
}